requirements.txt
alembic==1.4.2
amqp==2.6.0
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
blinker==1.4
bs4==0.0.1
celery==4.4.6
certifi==2020.6.20
click==7.1.2
cloudinary==1.21.1
DateTime==4.3
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
future==0.18.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathlib==1.0.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.6
pyotp==2.3.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.0
